+++ sum-it/Source/main/Dialog/RDialog.cpp
@@ -1,7 +1,7 @@
 /*
 	Copyright 1996, 1997, 1998, 2000
 	        Hekkelman Programmatuur B.V.  All rights reserved.
-	
+
 	Redistribution and use in source and binary forms, with or without
 	modification, are permitted provided that the following conditions are met:
 	1. Redistributions of source code must retain the above copyright notice,
@@ -11,13 +11,13 @@
 	   and/or other materials provided with the distribution.
 	3. All advertising materials mentioning features or use of this software
 	   must display the following acknowledgement:
-	   
+
 	    This product includes software developed by Hekkelman Programmatuur B.V.
-	
+
 	4. The name of Hekkelman Programmatuur B.V. may not be used to endorse or
 	   promote products derived from this software without specific prior
 	   written permission.
-	
+
 	THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES,
 	INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
 	FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL
@@ -27,11 +27,11 @@
 	OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 	WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
 	OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
-	ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
+	ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 /*
 	RDialog, a dialog build from a resource template
-	
+
 	Copyright 1997, Hekkelman Programmatuur
 */
 
@@ -74,12 +74,12 @@ static float gFactor = -1;
 static void ReadRect(BPositionIO& stream, BRect& r)
 {
 	short left, top, right, bottom;
-	
+
 	stream.Read(&left, sizeof(short));
 	stream.Read(&top, sizeof(short));
 	stream.Read(&right, sizeof(short));
 	stream.Read(&bottom, sizeof(short));
-	
+
 	r.left = gFactor * left;
 	r.top = gFactor * top;
 	r.right = gFactor * right;
@@ -98,21 +98,21 @@ CRDialog::~CRDialog()
 
 void CRDialog::ConstructFromTemplate(BPositionIO& buf)
 {
-	long token, l, x, size;
+	int32 token, l, x, size;
 	BRect r;
 	char name[256], label[256];
 	BView *v = fMainView, *nv = NULL;
-	
+
 	size_t pos = buf.Position();
 	buf.Seek(0, SEEK_END);
 	size = buf.Position();
 	buf.Seek(pos, SEEK_SET);
-	
+
 	do
 	{
 		l = 4;
 		buf.Read(&token, l);
-		
+
 		switch (token)
 		{
 			case 'btn ':
@@ -213,7 +213,7 @@ void CRDialog::ConstructFromTemplate(BPositionIO& buf)
 				ReadRect(buf, r);
 				ReadCString(buf, 255, name);
 				nv = AddTextView(r, name, v);
-				break;	
+				break;
 			case 'user':
 				ReadRect(buf, r);
 				ReadCString(buf, 255, name);
@@ -239,7 +239,7 @@ void CRDialog::InitValues(int resID, BPositionIO& buf,
 {
 	if (gFactor < 0)
 		gFactor = be_plain_font->Size() / 10.0;
-		
+
 	if (resID >= 0)
 	{
 		const void *p = gResourceManager.LoadDialog(resID);
@@ -249,23 +249,23 @@ void CRDialog::InitValues(int resID, BPositionIO& buf,
 	}
 
 	buf.Seek(0, SEEK_SET);
-	
-	int token;
-	long l;
-	
+
+	uint32 token;
+	ssize_t l;
+
 	l = 4;
 	buf.Read(&token, l);
-	
+
 	if (token == 'dlog')
 		type = B_TITLED_WINDOW;
 	else if (token == 'mdlg')
 		type = B_MODAL_WINDOW;
 	else
 		THROW((errDamagedResources));
-	
+
 	ReadRect(buf, frame);
 	ReadCString(buf, 255, name);
-	
+
 	if (owner)
 	{
 		BRect r = owner->Frame();
@@ -278,12 +278,12 @@ void CRDialog::InitValues(int resID, BPositionIO& buf,
 		frame.OffsetBy((s.Frame().Width() - frame.Width()) / 2,
 			(s.Frame().Height() - frame.Height()) / 3);
 	}
-	
+
 	flags = B_NOT_RESIZABLE | B_NOT_ZOOMABLE;
 } /* CRDialog::InitValues */
 
 BView* CRDialog::AddUserView(BRect r, const char *name,
-	ulong v, BView *parent)
+	uint32 v, BView *parent)
 {
 	THROW((errDamagedResources));
 	return NULL;
